Reformat signing page files
We decided to have all the signing pages as one 'template' file, and the
rest of the data will be filled in. I just did some general reformmating
(more details below) for the template html, JS, and CSS file.
NO FUNCTIONALITY WAS CHANGED!!!

- Fixed indentation in the HTML file.
- Camel-cased some element IDs.
- Added comments.
- Added placeholder elements for where the data taken from SQL will go.
- Created element variables in the initialize() function to improve
  readability.
- Rearranged the placement of elements in the CSS file; the specifications
  for each element appear in the order they appear in the HTML file.

N.B. The original files were not replaced; these added files are not
     accessible from the index page.