Convert JavaScript functions to JQuery
This transition specifically addresses the future issue of repositioning
the textarea element that users will sign in. Doing this in pure JS will
be much more difficult than in JQuery. Additionally, it only makes sense
to fully convert the preexisting JS to JQuery as well now instead of
later, when there's more lines of code.