Add draggable and resizable signing area
Upon further research, I found that a textarea can be easily replaced
with a div that has editable contents. This opened back up the
possibility of having attachments with the text.

When the user clicks the sign button, they can click anywhere on either
page to create a signing area that is both resizable and draggable, only
constrained by its parent (the signing page). A major downside, however,
is that the user cannot use the mouse to select the text inside the
signing area; they can only press shift and use the arrow keys. I'm
working on finding a workaround for this.